requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyodbc
python-dotenv
pydantic
requests
openai-mcp
